Extract playlist analytics to a separate class
